+++ deepinv/transform/__init__.py
@@ -4,3 +4,4 @@ from .scale import Scale
 from .base import Transform
 from .projective import Homography
 from .reflect import Reflect
+from .temporal import ShiftTime