+++ deepinv/transform/reflect.py
@@ -17,7 +17,7 @@ class Reflect(Transform):
     See :class:`deepinv.transform.Transform` for further details and examples.
 
     :param int, list[int] dim: axis or axes on which to randomly select axes to reflect.
-    :param n_trans: number of transformed versions generated per input image.
+    :param int n_trans: number of transformed versions generated per input image.
     :param torch.Generator rng: random number generator, if None, use torch.Generator(), defaults to None
     """